% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_app.R
\name{App}
\alias{App}
\title{The main App class that encapsulates the Shiny app}
\value{
An App object

A Shiny app object that can be run with shiny::runApp()
}
\description{
The main App class that encapsulates the Shiny app

The main App class that encapsulates the Shiny app
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-App-new}{\code{App$new()}}
\item \href{#method-App-run}{\code{App$run()}}
\item \href{#method-App-clone}{\code{App$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-new"></a>}}
\if{latex}{\out{\hypertarget{method-App-new}{}}}
\subsection{Method \code{new()}}{
Initialize an App
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$new(modules, config, translator = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modules}}{A list of R6Generators, result of sourcing files}

\item{\code{config}}{A list, usually from reading json, that contains the specification of the dashboard}

\item{\code{translator}}{A subclass of ElementalTranslator, if you want to work with your own translation library, uses shiny.i18n package if NULL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-run"></a>}}
\if{latex}{\out{\hypertarget{method-App-run}{}}}
\subsection{Method \code{run()}}{
Create a Shiny app that can be run
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-clone"></a>}}
\if{latex}{\out{\hypertarget{method-App-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
