% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_theme}
\alias{create_theme}
\title{Create a theme for the Shiny app}
\usage{
create_theme(theme = "shiny")
}
\arguments{
\item{theme}{Name of a preset theme (ses bslib::bootswatch_themes())}
}
\value{
A theme object that can be used by bslib
}
\description{
Customizes a couple of bootstrap properties
}
