% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_auth.R
\name{ElementalAuth}
\alias{ElementalAuth}
\title{The Auth interface class that any authentication module should subclass in order to work with the Elemental framework}
\description{
Subclass it to implement a specific authentication technology or SSO etc. for your dashboard
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ElementalAuth-clone}{\code{ElementalAuth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalAuth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalAuth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalAuth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
