% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_data.R
\name{ElementalData}
\alias{ElementalData}
\title{The Data interface class that any database connector module should subclass in order to work with the Elemental framework}
\description{
Subclass it use a specific database backend for your dashboard
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ElementalData-new}{\code{ElementalData$new()}}
\item \href{#method-ElementalData-clone}{\code{ElementalData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalData$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
