% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_data.R
\name{ElementalDataSqlite}
\alias{ElementalDataSqlite}
\title{A basic SQLite-based implementation of the Data interface class}
\description{
Use it for development but not for production purposes. SQLite is file-based so it will not properly work with multiple concurrent users
}
\section{Super class}{
\code{\link[elemental:ElementalData]{elemental::ElementalData}} -> \code{ElementalDataSqlite}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ElementalDataSqlite-new}{\code{ElementalDataSqlite$new()}}
\item \href{#method-ElementalDataSqlite-clone}{\code{ElementalDataSqlite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalDataSqlite-new"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalDataSqlite-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalDataSqlite$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalDataSqlite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalDataSqlite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalDataSqlite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
