% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_translator.R
\name{ElementalTranslator}
\alias{ElementalTranslator}
\title{The Translator interface class that any i18n module should subclass in order to work with the Elemental framework}
\value{
A list of all available languages

The currently selected target language

The translated text as a character vector

A span element with attributes so that it can be updated automatically

A reactive value with the translation
}
\description{
The Translator interface class that any i18n module should subclass in order to work with the Elemental framework

The Translator interface class that any i18n module should subclass in order to work with the Elemental framework
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ElementalTranslator-new}{\code{ElementalTranslator$new()}}
\item \href{#method-ElementalTranslator-init_js}{\code{ElementalTranslator$init_js()}}
\item \href{#method-ElementalTranslator-set_language}{\code{ElementalTranslator$set_language()}}
\item \href{#method-ElementalTranslator-get_all_languages}{\code{ElementalTranslator$get_all_languages()}}
\item \href{#method-ElementalTranslator-get_current_language}{\code{ElementalTranslator$get_current_language()}}
\item \href{#method-ElementalTranslator-static_t}{\code{ElementalTranslator$static_t()}}
\item \href{#method-ElementalTranslator-ui_t}{\code{ElementalTranslator$ui_t()}}
\item \href{#method-ElementalTranslator-reactive_t}{\code{ElementalTranslator$reactive_t()}}
\item \href{#method-ElementalTranslator-clone}{\code{ElementalTranslator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslator-new"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslator-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Translator object

Always create a subclass of ElementalTranslator that implements the necessary functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslator$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslator-init_js"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslator-init_js}{}}}
\subsection{Method \code{init_js()}}{
Implement an init_js function that has to be called in the main ui to initialize the UI bindings.
This is necessary to enable ui_t functionality
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslator$init_js()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslator-set_language"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslator-set_language}{}}}
\subsection{Method \code{set_language()}}{
Set the translation language

Should be a language for which a translation exists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslator$set_language(language)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{language}}{The language code, as supported by the translation file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslator-get_all_languages"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslator-get_all_languages}{}}}
\subsection{Method \code{get_all_languages()}}{
Get a list of all available languages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslator$get_all_languages()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslator-get_current_language"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslator-get_current_language}{}}}
\subsection{Method \code{get_current_language()}}{
Get the currently selected target language
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslator$get_current_language()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslator-static_t"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslator-static_t}{}}}
\subsection{Method \code{static_t()}}{
Static translate function

Translates a given text to the chosen target language.
This does not react to changes in target language.
Useful for modals etc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslator$static_t(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{The text to translate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslator-ui_t"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslator-ui_t}{}}}
\subsection{Method \code{ui_t()}}{
UI translate function

Translates a given text into a span element that will automatically react to changes in the target language
Can be used wherever span/html elements are allowed. Cannot be used as an attribute of an element.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslator$ui_t(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{The text to translate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslator-reactive_t"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslator-reactive_t}{}}}
\subsection{Method \code{reactive_t()}}{
Reactive translate function

When a span is not allowed but a static translation is not possible, you can use this reactive translation.
It returns a reactive value with the translation, and it will automatically update when the target language changes.
However, you can only use this inside a server function, not directly in the ui. This means you will need uiOutput() and renderUI() to work with this.
It also means the whole renderUI block will be re-run when the language changes. Be careful of expensive computations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslator$reactive_t(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{The text to translate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
