% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_translator.R
\name{ElementalTranslatorI18n}
\alias{ElementalTranslatorI18n}
\title{The default implementation of the Translator class that implements the shiny.i18n package}
\description{
The default implementation of the Translator class that implements the shiny.i18n package

The default implementation of the Translator class that implements the shiny.i18n package
}
\section{Super class}{
\code{\link[elemental:ElementalTranslator]{elemental::ElementalTranslator}} -> \code{ElementalTranslatorI18n}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ElementalTranslatorI18n-new}{\code{ElementalTranslatorI18n$new()}}
\item \href{#method-ElementalTranslatorI18n-init_js}{\code{ElementalTranslatorI18n$init_js()}}
\item \href{#method-ElementalTranslatorI18n-set_language}{\code{ElementalTranslatorI18n$set_language()}}
\item \href{#method-ElementalTranslatorI18n-get_all_languages}{\code{ElementalTranslatorI18n$get_all_languages()}}
\item \href{#method-ElementalTranslatorI18n-get_current_language}{\code{ElementalTranslatorI18n$get_current_language()}}
\item \href{#method-ElementalTranslatorI18n-static_t}{\code{ElementalTranslatorI18n$static_t()}}
\item \href{#method-ElementalTranslatorI18n-ui_t}{\code{ElementalTranslatorI18n$ui_t()}}
\item \href{#method-ElementalTranslatorI18n-reactive_t}{\code{ElementalTranslatorI18n$reactive_t()}}
\item \href{#method-ElementalTranslatorI18n-clone}{\code{ElementalTranslatorI18n$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslatorI18n-new"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslatorI18n-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslatorI18n$new(initial_target_language)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslatorI18n-init_js"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslatorI18n-init_js}{}}}
\subsection{Method \code{init_js()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslatorI18n$init_js()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslatorI18n-set_language"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslatorI18n-set_language}{}}}
\subsection{Method \code{set_language()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslatorI18n$set_language(language)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslatorI18n-get_all_languages"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslatorI18n-get_all_languages}{}}}
\subsection{Method \code{get_all_languages()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslatorI18n$get_all_languages()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslatorI18n-get_current_language"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslatorI18n-get_current_language}{}}}
\subsection{Method \code{get_current_language()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslatorI18n$get_current_language()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslatorI18n-static_t"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslatorI18n-static_t}{}}}
\subsection{Method \code{static_t()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslatorI18n$static_t(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslatorI18n-ui_t"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslatorI18n-ui_t}{}}}
\subsection{Method \code{ui_t()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslatorI18n$ui_t(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslatorI18n-reactive_t"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslatorI18n-reactive_t}{}}}
\subsection{Method \code{reactive_t()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslatorI18n$reactive_t(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementalTranslatorI18n-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ElementalTranslatorI18n-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementalTranslatorI18n$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
