% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{Module}
\alias{Module}
\title{Base class for all modules supported by the framework}
\description{
Base class for all modules supported by the framework
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Module-initialize}{\code{Module$new()}}
    \item \href{#method-Module-get_id}{\code{Module$get_id()}}
    \item \href{#method-Module-get_title}{\code{Module$get_title()}}
    \item \href{#method-Module-set_title}{\code{Module$set_title()}}
    \item \href{#method-Module-is_active}{\code{Module$is_active()}}
    \item \href{#method-Module-get_inputs}{\code{Module$get_inputs()}}
    \item \href{#method-Module-get_input}{\code{Module$get_input()}}
    \item \href{#method-Module-get_outputs}{\code{Module$get_outputs()}}
    \item \href{#method-Module-get_output}{\code{Module$get_output()}}
    \item \href{#method-Module-get_intro_tour}{\code{Module$get_intro_tour()}}
    \item \href{#method-Module-serialize}{\code{Module$serialize()}}
    \item \href{#method-Module-get_state}{\code{Module$get_state()}}
    \item \href{#method-Module-get_params}{\code{Module$get_params()}}
    \item \href{#method-Module-get_param}{\code{Module$get_param()}}
    \item \href{#method-Module-stateful}{\code{Module$stateful()}}
    \item \href{#method-Module-is_fullscreen}{\code{Module$is_fullscreen()}}
    \item \href{#method-Module-set_fullscreen}{\code{Module$set_fullscreen()}}
    \item \href{#method-Module-get_ui}{\code{Module$get_ui()}}
    \item \href{#method-Module-set_input}{\code{Module$set_input()}}
    \item \href{#method-Module-set_param}{\code{Module$set_param()}}
    \item \href{#method-Module-remove}{\code{Module$remove()}}
    \item \href{#method-Module-start_server}{\code{Module$start_server()}}
    \item \href{#method-Module-clone}{\code{Module$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Module-initialize}{}}}
\subsection{\code{Module$new()}}{
  Default constructor for Module

Do not call directly, but subclass Module instead
This constructor will be called automatically when no constructor is present in child class
If necessary, create a constructor in child class, then don't forget to call super$initialize()
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$new(id, title, globals, module_inputs, state)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{An id for this module, can be generated with generate_id()}
      \item{\code{title}}{A optional title for this module, uses classname if not provided}
      \item{\code{globals}}{A reference to a set of global, reactive, variables. Don't use these directly in your own modules}
      \item{\code{module_inputs}}{A names list of character vectors that map output variables of other modules to input variables for this module}
      \item{\code{state}}{A list of variables that describe the current state of the module. Don't put datasets in here, just simple variables.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A Module object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_id}{}}}
\subsection{\code{Module$get_id()}}{
  Get id
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_id()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The id of this module
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_title"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_title}{}}}
\subsection{\code{Module$get_title()}}{
  Get title
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_title()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The title of this module
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-set_title"></a>}}
\if{latex}{\out{\hypertarget{method-Module-set_title}{}}}
\subsection{\code{Module$set_title()}}{
  Set title
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$set_title(title)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{title}}{The new title for this module}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-is_active"></a>}}
\if{latex}{\out{\hypertarget{method-Module-is_active}{}}}
\subsection{\code{Module$is_active()}}{
  Is this module active
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$is_active()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    TRUE or FALSE to denote whether the server function of this module has been started yet
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_inputs}{}}}
\subsection{\code{Module$get_inputs()}}{
  Get names of inputs
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_inputs()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A charachter vector of names of the necessary input variables
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_input"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_input}{}}}
\subsection{\code{Module$get_input()}}{
  Get an input value
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_input(input_var)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{input_var}}{The name of the input variable}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The current value of this input variable
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_outputs"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_outputs}{}}}
\subsection{\code{Module$get_outputs()}}{
  Get names of outputs
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_outputs()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A character vector of names of the exported output variables
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_output"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_output}{}}}
\subsection{\code{Module$get_output()}}{
  Get an output value
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_output(output_var)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{output_var}}{The name of the output variable}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The current value of this output variable
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_intro_tour"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_intro_tour}{}}}
\subsection{\code{Module$get_intro_tour()}}{
  Get the content of the introtour

This function is called when clicking the introtour button
Subclasses should generally override this function to provide a specific introtour for that module,
  otherwise this generic message will appear saying that there is no introtour yet,
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_intro_tour()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A list of steps, as required by the rintrojs package
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-Module-serialize}{}}}
\subsection{\code{Module$serialize()}}{
  Serialize this module

When making changes to the modules section of the config, all modules provide a list of values that are enough to recreate the module next time
This does not contain state, so the current selection or current value of input fields are not serialized here, but are stored in state instead.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$serialize()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A list of values that can be transformed to JSON
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_state"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_state}{}}}
\subsection{\code{Module$get_state()}}{
  Get the state of this module

This is not the configuration of the module, just the current state, like selection, slider value etc.

Be aware that state and config need to align to prevent errors.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_state()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A named list of values that together describe the current state
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_params"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_params}{}}}
\subsection{\code{Module$get_params()}}{
  Get names of static parameters

Static parameters provide the user a way to setup a module differently. They can only be used inside a module, not between modules.
Parameters are part of the modules' config, not its state.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_params()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A character vector of names of static parameters
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_param"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_param}{}}}
\subsection{\code{Module$get_param()}}{
  Get a parameter value
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_param(param)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{param}}{Name of the parameter}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The current value of the static parameter
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-stateful"></a>}}
\if{latex}{\out{\hypertarget{method-Module-stateful}{}}}
\subsection{\code{Module$stateful()}}{
  Convenience function to make a value stateful in the UI definition of a module

Wrap a value, for instance a slider's value or a textbox's value inside stateful and provide a default value.
This function will retrieve a value of it exists or register the default value as a state value otherwise.
It will automatically serialize in that case.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$stateful(varname, defaultvalue)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{varname}}{Name of the state variable}
      \item{\code{defaultvalue}}{The default value in case it does not exist yet}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The stored state value or the default value in case it doesn't exist
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-is_fullscreen"></a>}}
\if{latex}{\out{\hypertarget{method-Module-is_fullscreen}{}}}
\subsection{\code{Module$is_fullscreen()}}{
  Is this module currently in fullscreen mode?
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$is_fullscreen()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A reactive value that you can observe() that is TRUE or FALSE
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-set_fullscreen"></a>}}
\if{latex}{\out{\hypertarget{method-Module-set_fullscreen}{}}}
\subsection{\code{Module$set_fullscreen()}}{
  Change the fullscreen value

When a user clicks the fullscreen button or closes the fullscreen mode, 
this function is automatically called from the Tile object to set this 
reactive value that can be observed in your subclass.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$set_fullscreen(value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{value}}{TRUE or FALSE}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-get_ui"></a>}}
\if{latex}{\out{\hypertarget{method-Module-get_ui}{}}}
\subsection{\code{Module$get_ui()}}{
  Public wrapper function to retrieve the private ui

This ensures each Module has the same interface. 
Override the standard private$ui() function in your subclass
Otherwise a default message will appear as the UI.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$get_ui()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The UI for this module, usually calls the ui from the child module
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-set_input"></a>}}
\if{latex}{\out{\hypertarget{method-Module-set_input}{}}}
\subsection{\code{Module$set_input()}}{
  Reactively change the definition of an input variable
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$set_input(input_var, input_path)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{input_var}}{The name of the input variable}
      \item{\code{input_path}}{The new path where this module will look for a value to associate with this input variable. 
Always consists of a 2-long character vector with a module id and output variable name.

This method is defined within the server function so it operates in a reactive context.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-set_param"></a>}}
\if{latex}{\out{\hypertarget{method-Module-set_param}{}}}
\subsection{\code{Module$set_param()}}{
  Reactively change the value of a parameter
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$set_param(param_name, param_value)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{param_name}}{The name of the parameter}
      \item{\code{param_value}}{The new value of this parameter

This method is defined within the server function so it operates in a reactive context.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-remove"></a>}}
\if{latex}{\out{\hypertarget{method-Module-remove}{}}}
\subsection{\code{Module$remove()}}{
  Remove this module

Tries to remove all content, observers etc. to prevent memory leakage.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$remove()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-start_server"></a>}}
\if{latex}{\out{\hypertarget{method-Module-start_server}{}}}
\subsection{\code{Module$start_server()}}{
  Start the server function of this module
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$start_server()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A reactiveValues object with the output variables, even though that is not used anymore to access the output variables
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Module-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Module-clone}{}}}
\subsection{\code{Module$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Module$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
